Mv: reorganize CSS
