+++ client/package.json
@@ -12,6 +12,7 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
+    "coverage": "npm test --- --coverage",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
