+++ client/src/Alarm.js
@@ -6,7 +6,7 @@
 // month          0-12
 // day of week    0-6 (Sun-Sat)
 
-import './Alarm.css'
+import './styles/Alarm.css'
 import PropTypes from 'prop-types'
 import React from 'react'
