+++ client/src/index.js
@@ -1,4 +1,4 @@
-import './index.css'
+import './styles/index.css'
 import * as serviceWorker from './serviceWorker'
 import App from './App'
 import React from 'react'
